Add corepack support (#803)
nodejs is generally now being distributed with corepack, which takes
over yarn and pnpm binaries and allows support for yarn 2+ without
needing it to be committed to repositories, instead a "packageManager"
package.json field will define the version used, and downloaded on
execution of yarn or pnpm.

By default yarn 1 will be used, the same as yarn npm package. Currently,
like node docker image, `corepack enable` needs to be run before package
installation (or `corepack yarn ...`) to get this behaviour.